fix(keychain): float->BigInt crash in token send + HMR provider exports
C1: SendTokenDrawer computed base units as amount * 10**decimals and passed the
result to BigInt(), so a fractional amount produced a non-integer float
(0.0045 * 10**18 = 4499999999999999.5) and threw 'cannot be converted to a
BigInt because it is not an integer' (seen in prod on /inventory/token). Parse
the raw decimal input string into base units exactly with bigint math
(parseTokenAmount), thread the input string through amount.tsx, and treat
invalid/over-precision input as no-amount. Adds a regression test (0.0045 @ 18).

C2: provider modules exported createContext()/hooks/constants alongside the
Provider, breaking Vite Fast Refresh and invalidating context identity on
hot-reload (cause of dev-only 'useX must be used within XProvider' errors).
Enable react-refresh/only-export-components (warn) and split the offenders
(starterpack, onchain-purchase, credit-purchase, upgrade, wallets) into
context/provider/hook modules. Targeted modules warning-clean; tsc, lint
(0 errors), format, and keychain tests (579 passing) green.
