+++ packages/keychain/src/components/connect/create/ChooseSignupMethodForm.tsx
@@ -15,7 +15,7 @@ import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { SignupButton } from "../buttons/signup-button";
 import { credentialToAuth } from "../types";
 import { useUsernameValidation } from "./useUsernameValidation";
-import { useWallets } from "@/hooks/wallets";
+import { useWallets } from "@/hooks/use-wallets";
 
 export const INITIAL_OPTIONS: AuthOption[] = ["sms", "webauthn"];
 export const WALLET_OPTIONS: AuthOption[] = [
