+++ packages/keychain/src/components/slot/crypto-fund.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { parseTokenAmount } from "./crypto-fund";
+import { parseTokenAmount } from "@/utils/token-amount";
 
 describe("parseTokenAmount", () => {
   it("parses whole token amounts", () => {
@@ -10,6 +10,7 @@ describe("parseTokenAmount", () => {
   it("parses fractional token amounts", () => {
     expect(parseTokenAmount("10.25", 6)).toBe(10_250_000n);
     expect(parseTokenAmount("0.000001", 6)).toBe(1n);
+    expect(parseTokenAmount("0.0045", 18)).toBe(4_500_000_000_000_000n);
   });
 
   it("rejects invalid precision", () => {
